2020/08/20200808_tensorflow_hub/requirements.txt
matplotlib~=3.3.0
tensorflow~=2.3.0
numpy~=1.18.5
Pillow~=7.2.0
tensorflow-hub~=0.8.0